+++ src/main/java/com/taskadapter/redmineapi/IssueManager.java
@@ -114,46 +114,6 @@ public class IssueManager {
     }
 
     /**
-     * DEPRECATED. use createIssue(Issue issue) instead.
-     * "projectKey" parameter required by this method duplicates what is already available in "issue" parameter.
-     * Also, some Redmine APIs require project database ID instead of the string key (e.g. when updating Issue).
-     * To keep things consistent, we propose using project database IDs for other operations as well, like
-     * for creating new issues. Which means you need to create a Project instance with the project numeric ID
-     * and set it to the new Issue object - see createIssue(Issue issue) method.
-     * <p>
-     * Sample usage:
-     * <pre>
-     * {@code
-     *   Issue issueToCreate = IssueFactory.create();
-     *   issueToCreate.setSubject("This is the summary line 123");
-     *   Issue newIssue = mgr.createIssue(PROJECT_KEY, issueToCreate);
-     * }
-     * </pre>
-     *
-     * @param projectKey The project "identifier". This is a string key like "project-ABC", NOT a database numeric ID.
-     * @param issue      the Issue object to create on the server.
-     * @return the newly created Issue.
-     * @throws RedmineAuthenticationException invalid or no API access key is used with the server, which
-     *                                 requires authorization. Check the constructor arguments.
-     * @throws NotFoundException       the project with the given projectKey is not found
-     * @throws RedmineException
-     * @see #createIssue(com.taskadapter.redmineapi.bean.Issue)
-     */
-    @Deprecated
-    public Issue createIssue(String projectKey, Issue issue) throws RedmineException {
-        final Project oldProject = issue.getProject();
-        final Project newProject = ProjectFactory.create();
-        newProject.setIdentifier(projectKey);
-        issue.setProject(newProject);
-        try {
-            return transport.addObject(issue, new BasicNameValuePair("include",
-                    Include.attachments.toString()));
-        } finally {
-            issue.setProject(oldProject);
-        }
-    }
-
-    /**
      * Sample usage:
      * <pre>
      * {@code